Added new unified frame types.
Unified           ID3v2.3 ID3v2.4 MP4  ASF                 Vorbis
---------------------------------------------------------------------------
Catalog Number                                             CATALOGNUMBER
Compilation       TCMP    TCMP    cpil                     COMPILATION
Encoding Settings TSSE    TSSE    ©too WM/EncodingSettings ENCODINGSETTINGS
Encoding Time             TDEN         WM/EncodingTime     ENCODINGTIME
Initial Key       TKEY    TKEY         WM/InitialKey       INITIALKEY
Mood                      TMOO         WM/Mood             MOOD
Release Country                                            RELEASECOUNTRY
Sort Album        TSOA    TSOA    soal WM/AlbumSortOrder   ALBUMSORT
Sort Album Artist TSO2    TSO2    soaa                     ALBUMARTISTSORT
Sort Artist       TSOP    TSOP    soar WM/ArtistSortOrder  ARTISTSORT
Sort Composer     TSOC    TSOC    soco                     COMPOSERSORT
Sort Name         TSOT    TSOT    sonm WM/TitleSortOrder   TITLESORT
WWW Audio File    WOAF    WOAF         WM/AudioFileURL     WWWAUDIOFILE
WWW Audio Source  WOAS    WOAS         WM/AudioSourceURL   WWWAUDIOSOURCE

+++ src/core/config/miscconfig.h
@@ -247,7 +247,7 @@ public:
   /** text encoding used for new ID3v2 tags */
   int m_textEncoding;
   /** frames which are displayed for Tag 2 even if not present */
-  quint32 m_quickAccessFrames;
+  quint64 m_quickAccessFrames;
   /** number of digits in track number */
   int m_trackNumberDigits;
   /** true to play file on double click */
